requirements.txt
flask>=2.0.0
gunicorn>=20.1.0
openai>=1.0.0
pinecone-client>=2.2.0
python-dotenv>=0.19.0
markdown>=3.4.0
httpx>=0.24.1
anthropic>=0.5.0
thefuzz>=0.19.0
python-Levenshtein>=0.21.1
rapidfuzz>=3.0.0
numpy>=1.24.0 